Move Inspector registration to Core project. Throw NotImplemented for multiple sites auditing function.
